- a couple of fixes like this one:
-    fprintf(f, "%g", real);
+    fprintf(f, "%g", static_cast<double>(real));

with 'real' being a fouble object. as fprintf is a c function with variable
arguments one cannot pass non-primitive objects through it and there is
no way for the compiler to figure out the right conversion operator, so
we give it a helping hand :)

+++ noncore/unsupported/qpdf/xpdf/Gfx.cc
@@ -1864,7 +1864,7 @@ void Gfx::opSetFont(Object args[], int numArgs) {
     printf("  font: tag=%s name='%s' %g\n",
 	   font->getTag()->getCString(),
 	   font->getName() ? font->getName()->getCString() : "???",
-	   args[1].getNum());
+	   static_cast<double>(args[1].getNum()));
     fflush(stdout);
   }
   state->setFont(font, args[1].getNum());
